requirements.txt
wheel==0.24.0
six==1.10.0
boto3==1.17.92
django==3.1.1
django-humanize==0.1.2
djangorestframework==3.11
django-data-browser==4.0.2
pytz==2021.1
lxml==4.5.2
PyPDF2==1.26.0
requests==2.12.1
fuzzywuzzy==0.18.0
django-rest-auth==0.8.2
jsonfield==1.0.3
kaleido==0.2.1
docutils==0.13.1
futures==3.1.1
matplotlib==3.7.1
mechanize==0.4.3
mysqlclient==2.0.1
python-dateutil==2.8.1
Unidecode==0.4.20
subprocess32==3.5.4
scipy==1.10.1
selenium==3.8.0
sentry-sdk==0.10.2
mock==4.0.3
authorizenet==1.1.4
pandas==1.5.2
numpy==1.24.1
django-import-export==2.4.0
django-storages==1.9.1
watchtower==1.0.6
tablib==3.0.0
django-extensions==3.1.1
tablib[ods]
odfpy==1.4.1
xlrd==2.0.1
xlwt==1.3.0
cairocffi==1.2.0
django-formtools==2.2
api2pdf==0.0.11
django-sql-explorer==2.1.3
django-data-browser==4.0.2
croniter==1.0.13
cron-descriptor==1.2.24
Pillow==7.2.0
plotly==5.15.0